fix components throwing error for ssr
